fix(signer): Fix `get_key` for KeyRequest::XOnlyPubkey
Previously we only accounted for `KeyRequest::Pubkey` when getting
the key for a SinglePriv. Since bitcoin 0.32.6, XOnlyPubkey is
available as a key request variant. To accommodate this the
implementation is changed to first construct a map of pubkey to
private key, and then call `get_key` on that.

Added unit test `signer::test::get_key_xonly_pk`.